[RLlib] Fix connector registry (#30635)
Introduce API to force re-register all the connectors in case the global registry is lost.
